+++ src/pages/myApps/components/myApp/index.tsx
@@ -3,6 +3,7 @@ import { Empty } from 'antd'
 import styles from './index.module.scss'
 import ApplicationCard from '@/components/ApplicationCard'
 import { useInstalledAppsStore } from '@/stores/installedApps'
+import { useApplicationCardModel } from '@/hooks/useApplicationCardModel'
 // import { useConfigStore } from '@/stores/appConfig'
 // import { uninstallApp } from '@/apis/invoke'
 
@@ -11,6 +12,7 @@ const MyApplications = () => {
     installedApps,
     // fetchInstalledApps,
   } = useInstalledAppsStore()
+  const { getCardState, handleInstall, uninstall } = useApplicationCardModel()
 
   // const { showBaseService } = useConfigStore()
   const [mergedApps, setMergedApps] = useState<API.INVOKE.InstalledApp[]>([])
@@ -107,11 +109,17 @@ const MyApplications = () => {
       {mergedApps.length > 0 ? <div className={styles.applicationList}>
         {
           mergedApps.map((item, index) => {
+            const cardState = getCardState(item)
             return (
               <ApplicationCard
                 key={`${item.appId}_${index}`}
                 appInfo={item}
                 operateId={0}
+                isInstalled={cardState.isInstalled}
+                hasUpdate={cardState.hasUpdate}
+                isInstalling={cardState.isInstalling}
+                onInstall={handleInstall}
+                onUninstall={uninstall}
               />
             )
           })